Split assets into directories.
JavaScript, CSS and templates are now split into directories for better
navigation in the IDE.

Renames all JavaScript library files to lowercase.
