+++ desktop/js/count_campaign_days.js
