+++ desktop/templates/application_of_funds.hbs
